requirements.txt
# Core & ML
torch==2.2.1
gymnasium[procgen]==0.29.1
numpy==1.26.4
pandas==2.0.3

# RL & curiosité
stable-baselines3==2.2.1
tensorboard==2.15.1
tqdm>=4.66.3,<5

# NLP & répétition espacée
transformers>=4.46.0
datasets>=3.0.0

# Web & visualisation
selenium==4.16.0
webdriver-manager==4.0.1
matplotlib==3.8.2
seaborn==0.13.1

# API temps réel
fastapi==0.109.2
uvicorn==0.27.1
websockets==12.0
pydantic==2.6.1

# Continual learning
accelerate>=0.22.0

# Dev & QA
black==24.1.1
flake8==7.0.0
scikit-learn==1.4.0
pytest==8.0.0
peft>=0.4.0

# Config parsing
pyyaml==6.0.1

# Extras for Gymnasium environments
requests>=2.32.2,<3
optuna==3.5.0
hydra-core==1.3.2
# RLHF methods (GRPOTrainer, etc.)
trl==0.17.0
bitsandbytes>=0.41.3  # main BitsAndBytes package for quantization support
deepspeed>=0.9.3
# Add other specific versions if needed

# Add setuptools and wheel for building packages from sdist
setuptools>=65.0.0
wheel>=0.38.4

# MPI for Python binding for distributed training
mpi4py==4.0.3

# JupyterLab
jupyterlab>=3.0

# Added from the code block
pdfplumber==0.9.0
PyPDF2==3.0.1
Pillow==10.0.0
